third_party/blink/web_tests/platform/linux/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/softsign.https.any_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] softsign positive float32 1D constant tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test softsign float32 actual 0.3117741644382477 should be close enough to expected 0.8030999898910522 by ULP distance: expected a number less than or equal to 3n but got 11400985n
[FAIL] softsign positive float16 1D constant tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test softsign float16 actual 0 should be close enough to expected 0.61279296875 by ULP distance: expected a number less than or equal to 3 but got 14567
[FAIL] softsign positive float16 0D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test softsign float16 actual 0 should be close enough to expected 0.61279296875 by ULP distance: expected a number less than or equal to 3 but got 14567
[FAIL] softsign negative float16 0D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test softsign float16 actual 0 should be close enough to expected -0.72216796875 by ULP distance: expected a number less than or equal to 3 but got 47559
[FAIL] softsign positive float16 1D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test softsign float16 actual -0.0078125 should be close enough to expected 0.61279296875 by ULP distance: expected a number less than or equal to 3 but got 26393
[FAIL] softsign negative float16 1D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test softsign float16 actual 0 should be close enough to expected -0.72216796875 by ULP distance: expected a number less than or equal to 3 but got 47559
[FAIL] softsign float16 2D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test softsign float16 actual 0 should be close enough to expected -0.89306640625 by ULP distance: expected a number less than or equal to 3 but got 47909
[FAIL] softsign float16 3D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test softsign float16 actual 0 should be close enough to expected -0.89306640625 by ULP distance: expected a number less than or equal to 3 but got 47909
[FAIL] softsign float16 4D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test softsign float16 actual 0 should be close enough to expected -0.89306640625 by ULP distance: expected a number less than or equal to 3 but got 47909
Harness: the test ran to completion.
